refactor: improved portal generate command and updated sdk version (#61)
What
This change refactors the complete portal:generate command flow, including the controller. In addition to that, the following things have been updated:

The APIMatic SDK has been updated to the latest version and is kept as part of the codebase (temporarily).
Updated the portal:generate command to use the SDK to call the Platform API instead of calling the Platform API's portal generation endpoint directly. Updated the error messaging corresponding to the error responses.
Added unit tests for the portal:generate command.
Updated some outdated dependencies to the latest versions.